minor update to nodes
